[AR] Make projectile shoot to where the crosshair/camera is looking.
This is done by sending a trace from the camera and making it collide
with the environment. We then save that position so it can be used by
the projectiles.

Also changed the naming conventions so that they're a bit closer to the
Google Style Guide. This might proove to be a bad idea in the long run.