[AR] Reorganized files so that include paths are easier to understand.
Each module will have a base directory of a the module name, which will
have the Build.cs file and the module loading files. There will be
another directory of the module name with the actual source, which will
be exposed as a Public include.

This permits other modules to include the code as:

`<module_name>/<header.h>`, which is better than just the header.