[Assignment 3] Base item interaction via spawner
The spawner now sends the request to the item to interact and
unspaws/respawns it as needed.

Changes need to be made so that the object can inform whether it can
interact with the caller (eg. a health item should not be used if the
user has full health, etc.).