[Lecture 14] Widget subsystem & Damage popup
This creates a new Widget Subsystem that is configured via a data table.
We also created a custom GameInstance class that can set the config data
table and configure the system.

This way we can load the asset on the beginning and use it via the game.
We can request widgets globally and get damage widgets. This can be done
via C++ (ARAICharacter) or via Blueprints (BP_TargetDummy).