[base] Make AR_DECLARE_DEBUG_CATEGORY unique across files.
The previous implementation had the problem that if two macros happened
at the same time, they would collide in their variable name.

With this approach, we add a name that should help make those global
variables unique. This also has the plus of making the console variable
a little nicer.