[interaction] Change interaction to work with collision volumes
Before it was working camera queries, which have real problems when
being replicated. The collision volumes should be more close to what is
expected to work when going into networking.

This also adds a new "Interaction" collision preset, which permits
objects to present a collision volume the character will detect.