[client_server_split] Moved to separate modules with correct defines
Now AR_BUILD_CLIENT and AR_BUILD_SERVER are being passed by UBT to
ensure that they are only defined in the correct modules regardless of
build type.

This means that a client module will not be able to include server code,
even when building a Game target.

Moved interaction to this new model.
Discovered that it's *VERY IMPORTANT* that UPROPERTY are not accessed
directly (having friend classes), since there are some |this| hacks made
by Unreal.

Solved by stopping making the client & server splits being friends of
the base, which forces it to go via accessors, which seems to make
Unreal happy.