[client_server_split] Correct per actor tracking of client/server.
Depending on the build type, we have an easier or harder way to
determine if an AActor or UActorComponent is running on the nomical
client or server. We now use this information to route the logic to the
correct split, even in the case when both are compiled in (like in the
case of Listen servers).

This currently works even with RUOP (Run Under Once Process), though
currently the heuristic to detect it is very simplistic (basically
assumes HasAuthority == Server), which will not be true for client only
actors. In that case, the logic will have to be augmented somehow.