[base] Created simple FSM framework
This permit to have states within one class.
It is not great as the flow is still embedded in the code rather than
being declarative, which is the main source of bugs of FSM over time.

But this is better than hand-rolled FSMs anyway.

The best approach is code generation, but that is a more involved
solution.