[console_commands] Initial Server Console Commands ground work
This will permit to define console commands that will only exist (and be
executed on the server).

The main idea is that console commands will be registered statically
(before main) on a map and this will be used by a game object to
populate the commands in the server upon initialization and make hooks
on the clients.