[attributes][ai] AI Minion now correctly rag-dolls on death.
Before the calculation was only done on the server. Now it is done in
both sides.

This also meant that the attribute calculation flags (eg. Killed) was
only done on the server, and not sent on the OnRep. This should be move
to a message, likely an RPC, to have more semantics rather than reverse
engineer them.